requirements.txt
pydub
webrtcvad
transformers==4.24.0
datasets==2.13.1
tokenizers==0.13.3
faster-whisper==0.9.0
tqdm
soundfile==0.12.1
kenlm==0.2.0
pyctcdecode==0.5.0
python-decouple==3.8
vietnam_number
urllib3==2.0.5
jiwer==3.0.2
argparse